+++ source/php/Notification/NotificationSenderInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace VolunteerManager\Notification;
+
+interface NotificationSenderInterface
+{
+    public function send(string $to, string $from, string $message): bool;
+}
